arreglo carpetas y fin generador coche
Falta generador personas
