+++ generador/README.md
