+++ generador/generators/pub_gen_ruta_persona.py
