+++ generador/generators/pub_generador_ruta_hardcode.py
