+++ generador/rutas/rutas_coche/Indicaciones de C de Menorca, 19, Camins al Grau, 46023 València, Valencia, España a C del Sant Calze, 5, Ciutat Vella, 46001 València, Valencia, España.kml
