+++ generador/rutas/rutas_coche/Indicaciones de C. del Dr. Rodríguez de la Fuente, 12, Rascanya, 46019 València, Valencia, España a C del Pintor Matarana, 16, Benicalap, 46025 València, Valencia, España.kml
