+++ generador/rutas/rutas_coche/Indicaciones de Nuevo Centro, Av. de Pius XII, 2, Local 167, Campanar, 46009 Valencia, España a Ayora, C dels Sants Just i Pastor, 104, Camins al Grau, 46022 Valencia, España.kml
