--- rutas/ruta_prueba_coche/Indicaciones de Pl. de l'Ajuntament, 21, Ciutat Vella, 46002 València, Valencia, España a Av. de Suècia, 8, El Pla del Real, 46010 València, Valencia, España.kml
@@ -1,225 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<kml xmlns="http://www.opengis.net/kml/2.2">
-  <Document>
-    <name><![CDATA[Indicaciones de Pl. de l'Ajuntament, 21, Ciutat Vella, 46002 València, Valencia, España a Av. de Suècia, 8, El Pla del Real, 46010 València, Valencia, España]]></name>
-    <Style id="icon-1899-DB4436-nodesc-normal">
-      <IconStyle>
-        <color>ff3644db</color>
-        <scale>1</scale>
-        <Icon>
-          <href>https://www.gstatic.com/mapspro/images/stock/503-wht-blank_maps.png</href>
-        </Icon>
-        <hotSpot x="32" xunits="pixels" y="64" yunits="insetPixels"/>
-      </IconStyle>
-      <LabelStyle>
-        <scale>0</scale>
-      </LabelStyle>
-      <BalloonStyle>
-        <text><![CDATA[<h3>$[name]</h3>]]></text>
-      </BalloonStyle>
-    </Style>
-    <Style id="icon-1899-DB4436-nodesc-highlight">
-      <IconStyle>
-        <color>ff3644db</color>
-        <scale>1</scale>
-        <Icon>
-          <href>https://www.gstatic.com/mapspro/images/stock/503-wht-blank_maps.png</href>
-        </Icon>
-        <hotSpot x="32" xunits="pixels" y="64" yunits="insetPixels"/>
-      </IconStyle>
-      <LabelStyle>
-        <scale>1</scale>
-      </LabelStyle>
-      <BalloonStyle>
-        <text><![CDATA[<h3>$[name]</h3>]]></text>
-      </BalloonStyle>
-    </Style>
-    <StyleMap id="icon-1899-DB4436-nodesc">
-      <Pair>
-        <key>normal</key>
-        <styleUrl>#icon-1899-DB4436-nodesc-normal</styleUrl>
-      </Pair>
-      <Pair>
-        <key>highlight</key>
-        <styleUrl>#icon-1899-DB4436-nodesc-highlight</styleUrl>
-      </Pair>
-    </StyleMap>
-    <Style id="line-1267FF-5000-nodesc-normal">
-      <LineStyle>
-        <color>ffff6712</color>
-        <width>5</width>
-      </LineStyle>
-      <BalloonStyle>
-        <text><![CDATA[<h3>$[name]</h3>]]></text>
-      </BalloonStyle>
-    </Style>
-    <Style id="line-1267FF-5000-nodesc-highlight">
-      <LineStyle>
-        <color>ffff6712</color>
-        <width>7.5</width>
-      </LineStyle>
-      <BalloonStyle>
-        <text><![CDATA[<h3>$[name]</h3>]]></text>
-      </BalloonStyle>
-    </Style>
-    <StyleMap id="line-1267FF-5000-nodesc">
-      <Pair>
-        <key>normal</key>
-        <styleUrl>#line-1267FF-5000-nodesc-normal</styleUrl>
-      </Pair>
-      <Pair>
-        <key>highlight</key>
-        <styleUrl>#line-1267FF-5000-nodesc-highlight</styleUrl>
-      </Pair>
-    </StyleMap>
-    <Placemark>
-      <name><![CDATA[Indicaciones de Pl. de l'Ajuntament, 21, Ciutat Vella, 46002 València, Valencia, España a Av. de Suècia, 8, El Pla del Real, 46010 València, Valencia, España]]></name>
-      <styleUrl>#line-1267FF-5000-nodesc</styleUrl>
-      <LineString>
-        <tessellate>1</tessellate>
-        <coordinates>
-          -0.37586,39.47036,0
-          -0.37583,39.47036,0
-          -0.37579,39.47037,0
-          -0.37571,39.47038,0
-          -0.37552,39.4704,0
-          -0.37542,39.4704,0
-          -0.37489,39.4704,0
-          -0.37463,39.47042,0
-          -0.37422,39.47043,0
-          -0.37421,39.47044,0
-          -0.37417,39.47044,0
-          -0.37409,39.47048,0
-          -0.37399,39.47052,0
-          -0.37377,39.47061,0
-          -0.37252,39.47124,0
-          -0.37216,39.47147,0
-          -0.37214,39.47147,0
-          -0.37202,39.47147,0
-          -0.37178,39.47145,0
-          -0.3711,39.4714,0
-          -0.3708,39.47136,0
-          -0.3705,39.47133,0
-          -0.37028,39.47168,0
-          -0.37021,39.47179,0
-          -0.37002,39.47222,0
-          -0.36955,39.47209,0
-          -0.36942,39.47205,0
-          -0.36917,39.47197,0
-          -0.36907,39.47194,0
-          -0.36905,39.47194,0
-          -0.36902,39.47194,0
-          -0.36899,39.47194,0
-          -0.36897,39.47194,0
-          -0.36895,39.47194,0
-          -0.36892,39.47195,0
-          -0.3689,39.47196,0
-          -0.36881,39.47199,0
-          -0.3688,39.47197,0
-          -0.36879,39.47196,0
-          -0.36878,39.47194,0
-          -0.36876,39.47193,0
-          -0.36875,39.47192,0
-          -0.36873,39.47191,0
-          -0.36872,39.4719,0
-          -0.3687,39.47189,0
-          -0.36865,39.47186,0
-          -0.3686,39.47185,0
-          -0.36855,39.47184,0
-          -0.3685,39.47183,0
-          -0.36845,39.47183,0
-          -0.3684,39.47184,0
-          -0.36835,39.47184,0
-          -0.36831,39.47186,0
-          -0.36819,39.47184,0
-          -0.36793,39.4718,0
-          -0.36774,39.47174,0
-          -0.36736,39.47154,0
-          -0.36675,39.47124,0
-          -0.36547,39.47061,0
-          -0.36529,39.47066,0
-          -0.36523,39.4707,0
-          -0.36517,39.47072,0
-          -0.36512,39.47075,0
-          -0.36433,39.4711,0
-          -0.36355,39.47144,0
-          -0.36345,39.47148,0
-          -0.36336,39.47152,0
-          -0.36333,39.47154,0
-          -0.36329,39.47155,0
-          -0.36319,39.4716,0
-          -0.36308,39.47167,0
-          -0.36297,39.47174,0
-          -0.36271,39.47186,0
-          -0.36157,39.47235,0
-          -0.3615,39.47238,0
-          -0.36147,39.47239,0
-          -0.36142,39.47242,0
-          -0.36041,39.47291,0
-          -0.36041,39.47299,0
-          -0.3604,39.47299,0
-          -0.36039,39.47299,0
-          -0.36038,39.47299,0
-          -0.36037,39.47299,0
-          -0.36036,39.47299,0
-          -0.36035,39.47299,0
-          -0.36034,39.47299,0
-          -0.36033,39.47299,0
-          -0.36032,39.47299,0
-          -0.36031,39.473,0
-          -0.3603,39.473,0
-          -0.36029,39.473,0
-          -0.36029,39.47301,0
-          -0.36028,39.47301,0
-          -0.36027,39.47302,0
-          -0.36026,39.47302,0
-          -0.36026,39.47303,0
-          -0.36025,39.47304,0
-          -0.36024,39.47304,0
-          -0.36024,39.47305,0
-          -0.36023,39.47305,0
-          -0.36023,39.47306,0
-          -0.36023,39.47307,0
-          -0.36022,39.47308,0
-          -0.36022,39.47309,0
-          -0.36022,39.4731,0
-          -0.36021,39.4731,0
-          -0.36021,39.47311,0
-          -0.36021,39.47313,0
-          -0.36005,39.47313,0
-          -0.36005,39.47314,0
-          -0.35975,39.47374,0
-          -0.35968,39.4739,0
-          -0.35967,39.47392,0
-          -0.35962,39.47403,0
-          -0.35955,39.47418,0
-          -0.35953,39.47422,0
-          -0.35951,39.47427,0
-          -0.35937,39.47455,0
-          -0.35936,39.47457,0
-          -0.35932,39.47466,0
-          -0.3592,39.47489,0
-          -0.3591,39.47507,0
-        </coordinates>
-      </LineString>
-    </Placemark>
-    <Placemark>
-      <name><![CDATA[Pl. de l'Ajuntament, 21, Ciutat Vella, 46002 València, Valencia, España]]></name>
-      <styleUrl>#icon-1899-DB4436-nodesc</styleUrl>
-      <Point>
-        <coordinates>
-          -0.3758577,39.4703598,0
-        </coordinates>
-      </Point>
-    </Placemark>
-    <Placemark>
-      <name>Av. de Suècia, 8, El Pla del Real, 46010 València, Valencia, España</name>
-      <styleUrl>#icon-1899-DB4436-nodesc</styleUrl>
-      <Point>
-        <coordinates>
-          -0.3591019,39.4750696,0
-        </coordinates>
-      </Point>
-    </Placemark>
-  </Document>
-</kml>
