--- rutas/ruta_prueba_peaton/Indicaciones de C de Sant Pau, 1, Ciutat Vella, 46002 València, Valencia, España a Pg. de l'Albereda, 21, El Pla del Real, 46010 València, Valencia, España.kml
@@ -1,161 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<kml xmlns="http://www.opengis.net/kml/2.2">
-  <Document>
-    <name><![CDATA[Indicaciones de C/ de Sant Pau, 1, Ciutat Vella, 46002 València, Valencia, España a Pg. de l'Albereda, 21, El Pla del Real, 46010 València, Valencia, España]]></name>
-    <Style id="icon-1899-DB4436-nodesc-normal">
-      <IconStyle>
-        <color>ff3644db</color>
-        <scale>1</scale>
-        <Icon>
-          <href>https://www.gstatic.com/mapspro/images/stock/503-wht-blank_maps.png</href>
-        </Icon>
-        <hotSpot x="32" xunits="pixels" y="64" yunits="insetPixels"/>
-      </IconStyle>
-      <LabelStyle>
-        <scale>0</scale>
-      </LabelStyle>
-      <BalloonStyle>
-        <text><![CDATA[<h3>$[name]</h3>]]></text>
-      </BalloonStyle>
-    </Style>
-    <Style id="icon-1899-DB4436-nodesc-highlight">
-      <IconStyle>
-        <color>ff3644db</color>
-        <scale>1</scale>
-        <Icon>
-          <href>https://www.gstatic.com/mapspro/images/stock/503-wht-blank_maps.png</href>
-        </Icon>
-        <hotSpot x="32" xunits="pixels" y="64" yunits="insetPixels"/>
-      </IconStyle>
-      <LabelStyle>
-        <scale>1</scale>
-      </LabelStyle>
-      <BalloonStyle>
-        <text><![CDATA[<h3>$[name]</h3>]]></text>
-      </BalloonStyle>
-    </Style>
-    <StyleMap id="icon-1899-DB4436-nodesc">
-      <Pair>
-        <key>normal</key>
-        <styleUrl>#icon-1899-DB4436-nodesc-normal</styleUrl>
-      </Pair>
-      <Pair>
-        <key>highlight</key>
-        <styleUrl>#icon-1899-DB4436-nodesc-highlight</styleUrl>
-      </Pair>
-    </StyleMap>
-    <Style id="line-1267FF-5000-nodesc-normal">
-      <LineStyle>
-        <color>ffff6712</color>
-        <width>5</width>
-      </LineStyle>
-      <BalloonStyle>
-        <text><![CDATA[<h3>$[name]</h3>]]></text>
-      </BalloonStyle>
-    </Style>
-    <Style id="line-1267FF-5000-nodesc-highlight">
-      <LineStyle>
-        <color>ffff6712</color>
-        <width>7.5</width>
-      </LineStyle>
-      <BalloonStyle>
-        <text><![CDATA[<h3>$[name]</h3>]]></text>
-      </BalloonStyle>
-    </Style>
-    <StyleMap id="line-1267FF-5000-nodesc">
-      <Pair>
-        <key>normal</key>
-        <styleUrl>#line-1267FF-5000-nodesc-normal</styleUrl>
-      </Pair>
-      <Pair>
-        <key>highlight</key>
-        <styleUrl>#line-1267FF-5000-nodesc-highlight</styleUrl>
-      </Pair>
-    </StyleMap>
-    <Placemark>
-      <name><![CDATA[Indicaciones de C/ de Sant Pau, 1, Ciutat Vella, 46002 València, Valencia, España a Pg. de l'Albereda, 21, El Pla del Real, 46010 València, Valencia, España]]></name>
-      <styleUrl>#line-1267FF-5000-nodesc</styleUrl>
-      <LineString>
-        <tessellate>1</tessellate>
-        <coordinates>
-          -0.37927,39.46897,0
-          -0.3792,39.46905,0
-          -0.37913,39.46915,0
-          -0.37896,39.46938,0
-          -0.37889,39.46947,0
-          -0.37882,39.46956,0
-          -0.37864,39.46981,0
-          -0.37805,39.46962,0
-          -0.37769,39.46951,0
-          -0.37684,39.46925,0
-          -0.3767,39.46923,0
-          -0.37609,39.46923,0
-          -0.37589,39.46917,0
-          -0.37579,39.46916,0
-          -0.3757,39.46915,0
-          -0.37554,39.46914,0
-          -0.3755,39.46915,0
-          -0.37483,39.46914,0
-          -0.37432,39.46915,0
-          -0.37408,39.46916,0
-          -0.3739,39.46916,0
-          -0.37234,39.46917,0
-          -0.37229,39.46917,0
-          -0.37224,39.46916,0
-          -0.37221,39.46916,0
-          -0.37218,39.46916,0
-          -0.37216,39.46915,0
-          -0.37211,39.46914,0
-          -0.37202,39.46909,0
-          -0.37178,39.46925,0
-          -0.37142,39.46951,0
-          -0.37112,39.46973,0
-          -0.37105,39.46979,0
-          -0.37037,39.47033,0
-          -0.37002,39.47026,0
-          -0.36855,39.47026,0
-          -0.36819,39.47025,0
-          -0.36705,39.47024,0
-          -0.36534,39.47021,0
-          -0.36523,39.47021,0
-          -0.36547,39.47061,0
-          -0.36529,39.47066,0
-          -0.36523,39.4707,0
-          -0.36517,39.47072,0
-          -0.36512,39.47075,0
-          -0.36433,39.4711,0
-          -0.36355,39.47144,0
-          -0.36345,39.47148,0
-          -0.36336,39.47152,0
-          -0.36333,39.47154,0
-          -0.36329,39.47155,0
-          -0.36319,39.4716,0
-          -0.36308,39.47167,0
-          -0.36297,39.47174,0
-          -0.36271,39.47186,0
-          -0.36294,39.4723,0
-          -0.36301,39.4724,0
-          -0.36311,39.47254,0
-        </coordinates>
-      </LineString>
-    </Placemark>
-    <Placemark>
-      <name>C/ de Sant Pau, 1, Ciutat Vella, 46002 València, Valencia, España</name>
-      <styleUrl>#icon-1899-DB4436-nodesc</styleUrl>
-      <Point>
-        <coordinates>
-          -0.3792666,39.4689663,0
-        </coordinates>
-      </Point>
-    </Placemark>
-    <Placemark>
-      <name><![CDATA[Pg. de l'Albereda, 21, El Pla del Real, 46010 València, Valencia, España]]></name>
-      <styleUrl>#icon-1899-DB4436-nodesc</styleUrl>
-      <Point>
-        <coordinates>
-          -0.3631121,39.4725412,0
-        </coordinates>
-      </Point>
-    </Placemark>
-  </Document>
-</kml>
